Add granular access control and approval mechanisms.
Introduced role-based access checks and approval workflows across backend and frontend. Backend supports `is_approved` flag and role-based filtering, while frontend applies visibility logic for actions (e.g., approvals, deletions) based on roles. Enhanced invitation and submission handling with role-defined constraints.

+++ backend/app/blueprints/objects_blueprint.py
@@ -5,7 +5,10 @@ from app.helpers.decorators import requires_user, validate_request_with, require
 from app.methods import get_objects_types, get_object_type_by_code, get_available_objects_by_type_code, create_object, \
     get_available_objects, get_object_by_id, update_object, delete_object, update_object_children, create_comment, \
     delete_comment, get_comment_by_id, create_submission, get_form_by_id, get_submission_by_id, update_submission, \
-    delete_submission, get_object_submissions
+    delete_submission, get_object_submissions, can_create_by_type, can_get_submission, approve_object, \
+    approve_submission
+from app.methods.access_methods import can_fill_in_category, can_modify_object, can_comment_object, can_delete_comment, \
+    can_modify_submission, can_delete_object, can_get_object_type, can_get_object
 from app.presenters.presenters import present_object, present_object_type, present_connected_object, present_comment, \
     present_submission
 from app.validators import validate_object, validate_object_children, validate_comment, validate_submission
@@ -16,81 +19,90 @@ objects_blueprint = Blueprint('objects', __name__, url_prefix='/objects')
 @objects_blueprint.route('/types', methods=['GET'])
 @requires_user
 def object_types(user):
-    return jsonify([present_object_type(object_type) for object_type in get_objects_types()])
+    return jsonify([present_object_type(object_type, user) for object_type in get_objects_types() if
+                    can_get_object_type(user, object_type)])
 
 
 @objects_blueprint.route('/<string:type_code>', methods=['GET'])
 @requires_user
 def object_type_endpoint(user, type_code):
     objects = get_available_objects_by_type_code(type_code)
-    return jsonify([present_object(obj) for obj in objects]), 200
+    return jsonify([present_object(obj) for obj in objects if can_get_object(user, obj)]), 200
 
 
 @objects_blueprint.route('', methods=['GET'])
 @requires_user
 def objects_endpoint(user):
     objects = get_available_objects()
-    return jsonify([present_object(obj) for obj in objects]), 200
+    return jsonify([present_object(obj) for obj in objects if can_get_object(user, obj)]), 200
 
 
 @objects_blueprint.route('/<string:type_code>/create', methods=['POST'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 @validate_request_with(validate_object)
 def create_object_endpoint(validated_data, user, type_code):
     object_type = get_object_type_by_code(type_code)
-    return jsonify(present_object(create_object(user, object_type, validated_data))), 201
+
+    if can_create_by_type(user, object_type):
+        return jsonify(present_object(create_object(user, object_type, validated_data))), 201
 
 
-from flask import request
-
+@objects_blueprint.route('/<int:object_id>/approve', methods=['POST'])
+@requires_user
+@requires_roles(['admin', 'teacher'])
+def approve_object_endpoint(user, object_id,):
+    obj = get_object_by_id(object_id)
+    return jsonify(present_object(approve_object(user, obj))), 200
 
 @objects_blueprint.route('/<int:object_id>', methods=['PUT'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 @validate_request_with(validate_object)
 def update_object_endpoint(validated_data, user, object_id):
     obj = get_object_by_id(object_id)
-    return jsonify(present_object(update_object(obj, validated_data))), 200
+    if can_modify_object(user, obj):
+        return jsonify(present_object(update_object(obj, validated_data))), 200
 
 
 @objects_blueprint.route('/<int:object_id>', methods=['DELETE'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 def delete_object_endpoint(user, object_id):
     obj = get_object_by_id(object_id)
-    delete_object(user, obj)
-    return jsonify({'deteted': True}), 200
+
+    if can_delete_object(user, obj):
+        delete_object(user, obj)
+        return jsonify({'deteted': True}), 200
 
 
 @objects_blueprint.route('/<int:object_id>/children', methods=['PUT'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 @validate_request_with(validate_object_children)
 def update_object_children_endpoint(validated_data, user, object_id):
     obj = get_object_by_id(object_id)
     children_ids = validated_data.get('children', [])
-    updated_object = update_object_children(obj, children_ids)
-    return jsonify([present_connected_object(child) for child in updated_object.children]), 200
+    if can_modify_object(user, obj):
+        updated_object = update_object_children(obj, children_ids)
+        return jsonify([present_connected_object(child) for child in updated_object.children]), 200
 
 
 @objects_blueprint.route('/<int:object_id>/comments', methods=['POST'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 @validate_request_with(validate_comment)
 def add_comment(validated_data, user, object_id):
     object = get_object_by_id(object_id)
-    return jsonify(present_comment(create_comment(user, object, validated_data))), 201
+
+    if can_comment_object(user, object):
+        return jsonify(present_comment(create_comment(user, object, validated_data))), 201
 
 
 @objects_blueprint.route('/<int:object_id>/comments/<int:comment_id>', methods=['DELETE'])
 @requires_user
-@requires_roles(['admin', 'teacher'])
 def delete_comment_endpoint(user, object_id, comment_id):
     object = get_object_by_id(object_id)
     comment = get_comment_by_id(comment_id)
-    delete_comment(user, comment)
-    return jsonify({'deteted': True}), 200
+
+    if can_delete_comment(user, comment):
+        delete_comment(user, comment)
+        return jsonify({'deteted': True}), 200
 
 
 @objects_blueprint.route('/<int:object_id>/forms/<int:form_id>', methods=['POST'])
@@ -99,15 +111,16 @@ def delete_comment_endpoint(user, object_id, comment_id):
 def create_submission_endpoint(validated_data, user, object_id, form_id):
     object = get_object_by_id(object_id)
     form = get_form_by_id(form_id)
-    new_sub = create_submission(user, form, object, validated_data)
-    return jsonify(present_submission(new_sub)), 201
+    if can_fill_in_category(user, form.category):
+        new_sub = create_submission(user, form, object, validated_data)
+        return jsonify(present_submission(new_sub)), 201
 
 
 @objects_blueprint.route('/<int:object_id>/submissions', methods=['GET'])
 @requires_user
 def get_submissions_endpoint(user, object_id):
     object = get_object_by_id(object_id)
-    return [present_submission(submission) for submission in get_object_submissions(object)], 200
+    return [present_submission(submission) for submission in get_object_submissions(object) if can_get_submission(user, submission)], 200
 
 
 @objects_blueprint.route('/<int:object_id>/submissions/<int:submission_id>', methods=['PUT'])
@@ -115,8 +128,17 @@ def get_submissions_endpoint(user, object_id):
 @validate_request_with(validate_submission)
 def update_submission_endpoint(validated_data, user, object_id, submission_id):
     sub = get_submission_by_id(submission_id)
-    updated = update_submission(sub, validated_data)
-    return jsonify(present_submission(updated)), 200
+
+    if can_modify_submission(user, sub):
+        updated = update_submission(sub, validated_data)
+        return jsonify(present_submission(updated)), 200
+
+@objects_blueprint.route('/<int:object_id>/submissions/<int:submission_id>/approve', methods=['POST'])
+@requires_user
+@requires_roles(['admin', 'teacher'])
+def approve_submission_endpoint(user, object_id, submission_id):
+    sub = get_submission_by_id(submission_id)
+    return jsonify(present_submission(approve_submission(user, sub))), 200
 
 
 @objects_blueprint.route('/<int:object_id>/submissions/<int:submission_id>', methods=['DELETE'])
@@ -124,5 +146,7 @@ def update_submission_endpoint(validated_data, user, object_id, submission_id):
 @requires_roles(['admin', 'teacher'])
 def delete_submission_endpoint(user, object_id, submission_id):
     sub = get_submission_by_id(submission_id)
-    delete_submission(user, sub)
-    return jsonify({'deleted': True}), 200
+
+    if can_modify_submission(user, sub):
+        delete_submission(user, sub)
+        return jsonify({'deleted': True}), 200
